FP.3 : Associate Keypoint Correspondences with Bounding Boxes + FP.4 : Compute Camera-based TTC
FP.3 : Associate Keypoint Correspondences with Bounding Boxes

Before a TTC estimate can be computed, we need to find all keypoint matches that belong to each 3D object. We can do this by simply checking whether the corresponding keypoints are within the region of interest in the camera image. All matches which satisfy this condition should be added to a vector. To eliminate outliers, we compute a robust mean of all the euclidean distances between keypoint matches and then remove those that are too far away from the mean.

FP.4 : Compute Camera-based TTC

Once keypoint matches have been added to the bounding boxes, the next step is to compute the TTC estimate.